back/drf_server/requirements.txt
asgiref==3.4.1
backcall==0.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
decorator==5.1.0
Django==3.2.8
django-cors-headers==3.7.0
django-extensions==3.1.3
django-seed==0.3.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
Faker==9.3.1
idna==3.3
ipython==7.28.0
jedi==0.18.0
jsonfield==3.1.0
matplotlib-inline==0.1.3
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.20
psycopg2==2.9.1
Pygments==2.10.0
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
text-unidecode==1.3
toposort==1.7
traitlets==5.1.0
urllib3==1.26.7
wcwidth==0.2.5
